Removed superfluous conditionals for a hex int literal
Some time ago, I removed the ability to submit hexadecimal integer literals with a capital "0X" (mainly because "0x" looks better, and there should only be one way to do it). This "policy" was enforced via the Lexer, but I removed it from the this section as well, because it's unnecessary.